HEALIX_HMS/healix-hms/api_gateway/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
python-jose[cryptography]
websockets
